requirements.txt
Django==4.2.24
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
mysqlclient==2.2.4
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.7.0
reportlab
requests
django-cors-headers
